images: fix 'gimp-debian-x64' build.
Apparently meson checks that the first value in 'command' is available and
executable. This is not the case for our GIMP (or gimp-console) executable which
is not installed at configuration time. So I hide the main executable name
behind 'sh' (which I assume will always be present, even on Windows these days?).

Fixes:

> ../subprojects/gimp-data/images/meson.build:18:2: ERROR: Program 'gimp-console-2.99' not found or not executable