images: generate desktop's application icons from the main logo SVG.
We now generate all the sizes which were committed as PNG from the desktop/
folder of the main repository, and even a bit more sizes (following Bruno Lopes'
example in !1154).

Even the smaller sizes (such as 16x16) are now generated from the SVG since we
simplified the logo enough so that Wilber is recognizable and we don't need
dedicated pixel-perfect versions anymore.