images: now generating gimp.ico from the new logo SVG.
The script is simplified since we can now generate all sizes from the same SVG
source.

This should also fix gimp#9483 (assuming GIMP's icon is taken from the .ico
file) since the problem was likely that because of the long brush of Wilber in
the previous icon, the head was scaled down and this is the likely reason why
GIMP's icon look too small next to other icons of the same size.

In her new version, Aryeom took special care that the (shorter) brush stops
exactly at the face's right border so that the logo can be cropped optimally all
around Wilber's face, without wasting a pixel.