icons: removing the gimp-wilber-outline icon.
I can clearly see it's not used anywhere anymore in code. As Symbolic variant,
it's not useful anyway since icons are recolored according to theme colors on
the fly. It might still be considered useful in the Color variant though, unless
we use shadow instead on the main logo which is more subtle than outlining the
icon and works very well even on a background exactly the same color as Wilber.