icons: use the shadow logo for gimp-wilber icons in Default icon theme.
The Legacy icon theme keeps the old Wilber. It's called "Legacy" for a reason
and some people might get nostalgic about our previous version, right?! :-D

The Default icon theme on the other hand will use the official logo in its color
variant, with an added shadow so that it displays well whatever the theme's
background color. This variant seems to work fine on any sizes, even smaller
ones.

The logo has also a unicolor variant (gimp-logo-unicolor.svg) which is going to
be used for the symbolic variant, except for sizes 32x32 and lower where Aryeom
made a simplified gimp-wilber-symbolic.svg to be used, where the mouth doesn't
look like weird details and the brush doesn't look like a dot which failed to be
cleaned out from the final drawing.