images: use the logo with shadow variant in About dialog.
For the desktop logo, we still use the base logo. In GNOME desktop at least, we
can see that GNOME Shell itself adds a shadow to our logo and therefore
providing a logo with a shadow already gets too much "shadowy".

I'm not sure about the .ico file used by Windows, but anyway I'm not sure that
Windows is themed too much and therefore requires to care about cases where one
would decide to theme their background color to the same flat gray as Wilber.

See for instance screenshot in #9483 where the menu shows some gradient and
therefore would not be a problem, whatever our choice of color.