images: remove clipPaths from all but the bottom.
A clip path was used for shadows in the canvas bottom, this is the only
relevant place. All the others were useless and were actually
interfering with display on some software, such as GNOME's image viewer
or web browsers.  Wilber was simply invisible when the SVG was displayed
on the web (at least on Gitlab).
Even though it was not a problem for GIMP (our script was still
generating the ICNS properly with Wilber in the canvas), that was a bit
annoying. It should now be fixed.

Additional note added later: we had this issue again later on, and after some
search, we found the relevant bug report in Inkscape:
https://gitlab.com/inkscape/inkscape/-/issues/1392

Bottom line: clipping paths should not contain groups. If such an issue happens
again of a design with clipping path becoming invisible in the browser, the
right solution is to release the clip, then ungroup the clip path, then re-clip.
The SVG will now be valid. Noting this in case this issue happens later so that
we don't have to search info on it again.