images: use DYLD_LIBRARY_PATH rather than LD_LIBRARY_PATH on macOS.
Note: I tried DYLD_FALLBACK_LIBRARY_PATH too, which works for the C plug-ins,
but not the Python plug-ins (it looks like GObject-Introspection does not use
this environment variable to search for libraries).

Anyway even though DYLD_LIBRARY_PATH may not be advised, neither is
LD_LIBRARY_PATH for normal usage. But here we are not in such normal usage. We
are just tricking the executable to find non-installed libraries for one-time
usage. So it's not really a problem.