images: generate and commit fileicon.ico.
I am now generating fileicon.ico from a base SVG representing a "file" and the
logo with shadow. Note that this file icon was not designed by Aryeom by lack of
time (I just reused a SVG file submitted in the tracker for this purpose, see
!1154) neither is the positioning of Wilber within this file design. It's just a
first version.

Similarly to gimp.ico, this file is not built by default and I commit the
generated '.ico' file because of a circular dependency (this file is named in
the gimp.rc which is compiled in as resource on Windows so it would break the
build if the file is absent.

This is why I also add the file update as a procedure step when updating the
logo.