images: updating the gimp_exe binary to a wrapping script.
In order to fix local builds for macOS, I am wrapping our locally built
executable in a shell script which will additionally update the rpath on macOS.
See gimp!1457.

A consequence of this is that we need to add a dependency to the built
executable (`gimp_exe` used to be it, but now it's just a script in the source
repository, so we add a `gimp_exe_depends`).